+++ exonum/src/blockchain/service.rs
@@ -184,7 +184,7 @@ pub trait Service: Send + Sync + 'static {
     /// `execute` for the service with the smallest ID is invoked first up to the largest one.
     /// Effectively, this means that services should not rely on a particular ordering of
     /// Service::execute invocations.
-    fn execute(&self, fork: &mut Fork) {}
+    fn execute_before_commit(&self, fork: &mut Fork) {}
 
     /// Handles block commit. This handler is invoked for each service after commit of the block.
     /// For example, a service can create one or more transactions if a specific condition